LPS-96808 Add a webpack registry module to synchronize portlet startup
We need an intermediate module to register webpack-based portlets that
are launched through the portlet extender. This is because the webpack
bundles are loaded through a <script> tag and thus started by the
browser in a separate lifecycle from that of the portlet extender (which
is driven by the Loader).

So, because we cannot assure which one will arrive first (if the browser
or the portlet extender) we need to store portlet extender launchings in
case the webpack bundle is not yet present. This is what this module is
for.

In addition, this module checks if more than one webpack-based portlet
has been placed in the same page, notifying the user in that case.

Note that we are using the bridge generator, to ease requiring this
module. Without the bridge we would need to care for the module's
version but that doesn't give any benefit because this module is
provided by the portal and, thus, there's no point in versioning it,
because there will always be only one instance of it and we make sure
that it is the right one when we set the constraints for the portlet
extender version in the MANIFEST files.