DO NOT MERGE WITHOUT REVISION!!! LPS-114977 Tie CSS variables to LayoutSet instead of Theme (as it is configured at that level)
Note that we are still storing the configured variables in company settings though it is not necessary, as we identify the settings with the LayoutSet id.

As it is not yet clear how we are going to store the CSS variables I'll leave it like this for now since it doesn't hurt, but we need to define this before releasing to the general public.