[WIP] LPS-116693 Support for require() in fragments
This commit simply checks fragments for a specific header with mappings from variable names to (NPM module, version constraints) pairs. When found, it wraps the fragment in a Liferay.Loader.require() call so that the variables are defined. This way, we may use any NPM module we like in fragments as long as it is listed in the header (which, usually, will be informed by the bundler).

Note that we assume that the header comes in the JS code (because the bundler has put it there). This is only for the PoC and should be moved to a more suitable place in the final version (perhaps the fragment.json file or something like that).