LPS-116693 Render React fragments (proof of concept commit)
This commit renders react fragments which have been bundled by https://github.com/liferay/liferay-js-toolkit/pull/627.

Note that there are two TODOs in this commit for yet unresolved needs:

1) Maintain fragments' JSModules updated with fragment's JS code (stored in DB)
2) Detect React vs non-React fragments

Requirement 1 is half-done lazily on demand in this PR, but the solution is
inefficient and incomplete, so it would be better to update JSModules in every
service where a FragmentEntry is modified/created/deleted. However we'd better
check with Echo team if there's a better option.

Requirement 2 maybe needs an extra field in the DB to store the fragment type.
We should check with Echo team too, to see what they think.

Other than that, the PR can be considered final.

Note, also, that we attach a new package.json. This is needed so that NPMRegistry
is aware of the `fragment-impl` JSBundle so that we can attach fragments'
JSModules to it.