LPS-127432 Remove methods to output resources from npm packages
We could also deprecate them, but we cannot provide any compatible implementation
so I guess the result would be the same. In fact I prefer the build to fail
rather than throwing at runtime.

Also, even if this is exposed through OSGi it does look like an internal API so
I think it's fair to do the breaking change.