LPS-126038 Fix I18N bug
Note that it does not make sense to use the Language.properties from our bundle,
but the one from the bundle containing the ClayDataSetDisplayView object's class.

This is because we use that ResourceBundle to localize the labels returned by the
ClayDataSetDisplayView, so it looks natural to get their values from the same
bundle.