LPS-130411 Adapt to DatasetDataSerializer API changes
We are removing the tableName parameter from the JAXRS API until is is proved
necessary. This is because after reviewing the code we have found that it
always uses the same data provider key for both the data provider and the
actions provider.

In case we want to differentiate in the future (which is quite probable) we
will re-add the parameter again, but giving it a proper name and, possibly,
refactoring the JAX API a bit.