LPS-136001 Enlarge url field
Set it to 1024 instead of 75 so that we may deploy these files inside document
library.

We may need to enlarge it more in the future, since there's no official max
length for an URL. For example, in the AMD loader we use 2,000 chars, but
proxies and web servers may use other values.

In any case, since this is an URL intended to be consumed as API, I don't expect
huge URLs here, so 1024 should suffice.