LPS-142295 Create sample Vanilla Counter remote app on DB creation
We create a sample row pointing to our Vanilla Counter app per company,
but only when the table is empty (ie: on fresh installations).

Note that this will recreate the sample if it is deleted, no other
remote app is created and portal is restarted. PM thinks this is a fair
behaviour but, if we see it creates any problem, we may change it in
the future.