LPS-142295 Add new event to PortalInstanceLifecycleListener
We need this event to detect when the Company has been added to the DB and its
portal instance (along with its ServletContext) has been started.

The presence of the ServletContext is ABSOLUTELY necessary to be able to
register portlets associated to the Company, so we cannot rely on the former
portalInstanceRegistered event if we need to register portlets in a recently
created Company.