<2,3> LPS-143013 Render import maps based on javascript portlet bundle's content
This is point 2 of the document. It allows deploying a project depending on an ESM-ready
npm package that is consumed as a bare identifier from the javascript code and
dereferenced to the node_modules directory by the import map.

An import map has the following appearance:

<script type="importmap">
{
  "scopes": {
    "/o/LPS-143013-test-1-1.0.0/": {
      "@esm-bundle/moment": "/o/LPS-143013-test-1-1.0.0/./node_modules/@esm-bundle/moment/esm/index.js"
    }
  }
}
</script>

It is rendered inside the <head> node and must be placed before any <script type="module">
is bootstrapped or the browser will complain.

----

For now we will be using a shim to honor import maps in Firefox. However it's not a
production ready solution because it shows fake error messages in the console. I suspect
they appear when the load fails, then the shim takes over and makes it work, but the
console error remains.

----

We are reading the importmap.json file from META-INF/resources inside the JAR. This is
because we don't have anything ready (yet) in the JS Toolkit to place it in a better
place but, in the final solution, we should probably move it to META-INF or the root
of the JAR file.