LPS-142295 New approach that DOES NOT WORK
This is how it should be done according to https://github.com/brianchandotcom/liferay-portal/pull/111290#issuecomment-1002313778

However, because RemoteAppEntry is an exotic entity, it does not work. More specifically,
everything works, but the portlet fails to register because of the Company and
Portal Instance lifecycle: right after the company is created, we insert the sample
custom element (we don't have any other moment in the lifecycle in which to insert it),
but because the virtual instance is not yet ready, the portlet fails to register.

So everything seems correct, from the point of view of Remote Apps administration, but
you cannot add the sample portlet to the page.