LPS-151395 Don't use internal paths of frontend-js-web
As we agreed, we are exporting the whole API through index.js, so don't
import internal paths because they break.

This is true for bundler2 bridges too, as they used to bridge internal
paths, so we will remove them as they are no longer needed.

Note that this may break external legacy code, so we may decide to leave
them for a while, or reintroduce them as needed.