LPS-151395 Add a new API to register scripts to be rendered at the bottom of the page
This is the equivalent to ScriptData in portal-kernel, but for the new
OSGi and ESM worlds.

Modifying ScriptData is very costly because it is inside portal-kernel (API
and implementation) so any little change usually causes a major version
increase. Thus, I'm leveraging this paradigm change in how scripts will be
handled to write this alternative.