LPS-155410 Create a marker interface for CETs with Properties
Not all CTEs use the Properties field, and we should expect that, if more
fields outside of the typeSettings appear, not all of them will be editable
for every CET, so I'm devising a new facet oriented framework in this
commit to be able to distinguish CET capabilities based on their type.

This way we can use instanceof checks from the UI implementation (or any
other place) to check wheter a given CET instance uses one of the DB fields
or not.