LPS-155911 Isolate AUI and AMD blocks of code in IIFE
I tried to isolate them in JS scope blocks {}, but that doesn't work when
multiple fragments refer to the same identifier and declare it with `var`,
so I'm going back to the old way of doing the isolation for AUI and AMD
fragments of code.

For ES fragments of code we'll begin to use {} since they have different
semantics (and should use const ot let instead of var, too).

Unfortunately, we cannot force that change for existing code because it
would break a lot of things and potentially customer code.