LPS-155911 Add a new globalDeclaration attribute to <aui:script>
This is a breaking change. It will be used to split <aui:script> code (with
no use, require, or load attribute) into executable chunks and declaration
chunks.

That is necessary because we must make sure all executable code belonging to a
portlet is executed atomically so that all fragments are synchronized and
executed in the order in which they appear.

Due to that and, because we are now using <script type=module>, there's no way
-now- to put things in global namespace (ie: window) so we need to provide one
way to do it if we want to support legacy code in <aui:script> blocks.

What the new mechanism is doing is placing all code in raw (with no deps)
<aui:script globalDeclaration=true> blocks in a separate
<script type=text/javascript> node so that it can define things in global
namespace.

Then, we need to make sure that we move executable code out of that node and put
it in the <script type=module> one, so that its execution is correctly
synchronized. Thus, we break some <aui:script> block into two new
blocks: <aui:script globalDeclaration=true> and <aui:script>.

At first I thought setting globalDeclaration's default value to true
would make things easier because most of the existing legacy code would
keep working even if the executable code was moved to the declaration
node. But then, I disliked a lot the resulting JSP code and realized
that it could lead to subtle errors which would be much harder to debug
than if some <aui:script> node was not annotated with
globalDeclaration=true when it should. So I decided to value easiness of
fixing/future maintenance over easiness of migrating legacy code.