LPS-165022 Render <script type=module> as <script type=module-shim> when es-module-shims is active
This is to use the "use es shim" switch to decide whether to use <script type=module> or
<script type=module-shim> in the whole portal.

This is because if you turn on the shims and use shim mode, you can/should route everything through
the shim, instead of the native loader. That way everything stays consistent and behaviour is
predictable.

So, the idea is:

  - Set the shims to off by default
  - If the user turns on the shims, everything is shimmed no matter what browser

Additionally, we will make liferay-cli use relative paths to obtain react instead of bare imports.

In the end, we (Liferay) don't need import maps at all, because we know where things are
(their URLs). Thus, only if a customer is interested in using import maps there will be
code with bare imports and, in that case, it's up to the customer deciding whether to use
the shims depending on their target browsers.