LPS-169939 Register the JSPackage client extension type with blade
This makes it possible to deploy a jsPackage client extension.

To use this it is necessary to run "gradle install" in the
gradle-plugins-workspace to make it copy the produced JAR to the .m2 folder
of your liferay-portal project.

Once the file is there, you just need to update the settings.gradle of your
blade workspace to point to liferay-portal's/.m2 repo, like this:

	repositories {
		maven {
			url "/home/you/liferay-portal/.m2"
		}

		...
	}

Then set the gradle-plugins-workspace version to 4.1.6, like this:

	classpath(group: "com.liferay", name: "com.liferay.gradle.plugins.workspace", version: "4.1.6") ...

And that will make gradle fetch the updated JAR file from your
liferay-portal/.m2 repo the next time you build any project in the
Workspace.

To create a client extension of the new type, you just need to add the
following fields to the client-extension.yaml file:

<project name>:
    identifier: <npm bare identifier>
    name: <client extension name>
    type: jsPackage
    url: <relative path to entry point .js file, eg: 'lib/index.js'>