LPS-170945 Improve performance in generation of URLs for module resolution servlet
This commit implements a distributed protocol between master and slave nodes of
a cluster to maintain a common random UUID as the /js_resolve_modules suffix

Any redeployment or restart event in any node causes a notification to the
cluster's master node that updates the master hash UUID and propagates it through
the slaves.

Since it is impossible to guarantee that this mechanism don't fails under some
corner scenarios (eg: if the frontend-js-loader-modules bundle is stopped or
missing from the master node), each node maintains a local UUID for those cases.

That way, the /js_resolve_modules endpoint always works for each node and the
worst thing that can happen if the cluster nodes fail to synchronize is that there
are more cache misses than could be. But the application remains stable and
nothing fails.

I have added exhaustive logging just in case we need it for support actions in the
future, if some situation we haven't thought of shows up.