LPS-170945 Make ResolveModulesHashUtil thread safe
This synchronization guarantees that the listeners are all given the correct
hash value and in the correct notification order so that the state of the
listeners is always consistent.

Not 100% sure about this, but I would say it shouldn't hurt performance too much,
since the synchronized methods are only invoked under rare circumstances and
they involve network communications too, so the cost of networking will always be
disproportionally big compared to the synchronization.

Also, network notifications (from master to slaves and vice-versa) are
fire-and-forget, so there won't be any network bottlenecks.