LPS-170945 Implement notifications synchronization
This accounts for the case when notifications from master to slaves arrive at
the wrong order at any slave. If an older notification arrives at a slave it is
ignored.

Notifications from slaves to master don't pose any problem because they don't
carry any information. In fact, burst notifications from slaves to master could
be coalesced and counted as just one (though we are not implementing that given
how unlikely it is and how the complexity of the solution would increase).

This algorithm guarantees that slaves have always the latest hash (or an older
one if one notification from master to slaves is lost), but in the case where
notifications arrive in the wrong order, the hash remains correct.

There's one case not covered enough, though: when a slave node becomes a master.
If this happens in the middle of a master-slaves notification or if a slave node
with an expired _lastSeenPriority becomes master, some of the updates may be
ignored and the cluster would be left in an inconsistent state.

However, once again, this shouldn't be a problem from the functional POV since
everything will keep working as usual. It's only that network caching will suffer
when/if browsers are moved between cluster nodes.