LPS-174464 Internalize resolutions hash computation in NPMRegistry
As per https://github.com/brianchandotcom/liferay-portal/pull/129875#issuecomment-1422389958

Note that we have changed NPMRegistryResolutionStateDigestUtil to receive
the list of modules and packages explicitly, instead of the NPMRegistry.
This guarantees that the digest contained in JSModulesCache will always be
coherent with the rest of the values in JSModulesCache, even if lazily
calculated.

I can't see any synchronization issue with this commit other than some
possible overhead in getResolutionStateDigest() if more than one thread
calculates the digest. But that's better than synchronizing the reads of
_resolutionStateDigest, since generally there will only be one write but
lots of reads from that variable.