LPS-174464 Add a getter to obtain a snapshot of the NPMRegistry
This will allow us to obtain stable information for a given hash no matter
how many times we invoke the NPMRegistry and even if it gets updated in
between them.

Note that, if the NPMRegistry is updated, the returned resolution may fail
in the browser (because it will try to access back JS modules from a state
of the NPMRegistry that doesn't exist any more) but the contents of the CDN
or proxy caches will be correct so that, in the future, in case the portal
goes back to the state when the resolution was computed, the CDN/proxy
caches are not posisoned by an incorrect resolution.

Going back to a previous state shouldn't be unusual: think of the case when
something is deployed, then undeployed after testing it.