WIP: LPS-176053 Incomplete migration of frontend-data-set-web to ESM
I'm migrating the things I need from AMD to ESM. This is because I need to
import the CE using ES and that is only possible by using the ESM build of
npm-scripts.

Since npm-scripts is generating ESM->AMD bridges, there's no need to
migrate the whole bundle, just some bits. But keep in mind that, in the
final PR, it should be fully migrated.