WIP: LPS-176053 PoC for Liferay APIs
This commit shows how we would write `Liferay` APIs inside this modules,
making having scattered snippets in a myriad of JSP files not necessary any
more, since with a single Java implementation and an interface, we will be
able to generate the .d.ts files, and implement both the server side of the
API, the `Liferay` implementation, and its wrapper.

This commit is not functional, it's only added for demonstration purposes.