LPS-176053 Define the data-set API
This shows how we will define public APIs so that they can be consumed by
our customers.

Note that, ideally, frontend-data-set-web should also build linting against
this API, but I'm not including that in this PR because I don't want to
migrate the full project to typescript now. However, we will need to do it
in the final PR.