LPS-183047 Scope portlets by company
This commit means that from now on, a portletName/Id is not enough to
address a certain portlet, because we need to know the companyId too.

This has several implications because it can be a breaking change depending
on how we change the consumer code.

In this commit we are inferring the companyId from the context, whether
because it is available directly or because we can extract it from the
ServiceContext or Company thread locals.

A different approach would be registering all portlets without the
com.liferay.company property under the SYSTEM company, as before, then
making portlets with com.liferay.company invisible unless new extended
APIs (to be defined) are used. This wouldn't be a breaking change, but the
implementation may be more complex.