LPS-183047 Fix friendly URL mappers
Since friendly URL mappers are assigned to portlets based on the portlet
name we need to check the companyId for company-scoped portlets to work.
Otherwise, if a duplicated portlet name is used for more than one company,
all URL mappers will end up in only one of them no matter which company
they are used for.

Note that this fix has been made because, after testing URL mappings of
custom elements it has been found that this was failing but, in general,
there may be more errors like this one spread across the whole portal,
which are due to the new assumption that portlets are scoped by company.

We don't know how many of these there are and how problematic they may be
to fix thus, we are going to try a new approach to scoping portlets by
company, which is a derivative of this PR, and see if we can do better.