LPS-183047 Refactor PanelApps to account for portlet model company scoping
We need to assume the SYSTEM company id for all panel apps because we don't
have any valid way to retrieve it.

This is most probably due to panel apps being deployed through code which
means they must be unique for all companies.

However, if we decide to scope panel apps by company in the future (imagine
defining a panel app CX, for instance), this code will break unless we make
it multitenant ready.
code may break.