LPS-183047 Refactor places to account for portlet model company scoping
We assume we can get the company id from the ServiceContextThreadLocal in a
variety of places. This is because we cannot get it from any other place.

Note that this code may fail if we invoke these methods from outside the
context of an HTTP request.

For the case of ResourceActionsImpl we must also look into the
CompanyThreadLocal because that's the thread local of choice when it comes
to database upgrades/validators.