LPS-183047 Enforce scoping by company for portlet models
Even though portlet models have always been scoped by company, the fact
that the implementations weren't made the code drift to the point that some
methods related to portlets weren't asking for the company id.

We restore and enfore that restriction again in the portlet services so
that nobody can use a portlet name or id alone to refer to a portlet.

# Conflicts:
#	portal-kernel/src/com/liferay/portal/kernel/service/PortletLocalServiceWrapper.java