LPS-188020 Create an import maps contributor to resolve @liferay/lang/...
The new I18N mechanism for ES modules relies on every module needing I18N
support importing an ES module named `@liferay/lang/<web context name>.js`.

This module is resolved to a URL containing the user's language id so that
the browser knows how to cache it and where to retrieve it.

The URL is pointing to a servlet that creates a JS that injects the
translations to the `Liferay.Language._map` on the fly.