LPS-188020 Servlet to inject translations to `Liferay.Language._map`
This servlet generates a dynamic JS file that injects the translations
needed by an OSGi bundle into `Liferay.Language._map`.

The list of language keys is taken from a well known file that is stored
under the `__liferay__/lang.json` in the same servlet context of the OSGi
bundle.

Those keys are translated to the language requested in the URL and a JS to
inject them to `Liferay.Language._map` is emitted so that it can be
executed in the browser.

Note that we are using the ETag header of HTTP to save bandwidth. However,
we don't store cached copies of the resources but only the ETags, to save
RAM and don't make code more complex (because we would need to manage the
cache to prevent it from growing a lot).

This limitation implies that we need to recompute the JS file for every
cache miss but, since the computational effort is so small, I don't think
it makes sense to make this caching more efficient for now. It could even
happen that managing a cache with the resources would be less performant
than recomputing the resource again.