LPS-178066 Don't add nonces for scripts in the CSP filter
This was plain wrong because it did a blind search/replace without parsing the HTML
which can lead to errors if the string "nonce" is used anywhere outside of a
<script> tag.

Also, this method is quite unsafe because it blesses any possible injected <script>
node that has not been emitted on purpose by DXP (think of some unescaped content
retrieved from the DB that was not intended to be executed, for example).

In other words, blindly adding nonces in a filter defeats the very purpose of the
CSP specification.

In following PRs we will be removing nonces for link and style.