LPS-195636 Simplify contract between FDS and filters
Remove the need to pass the id back since the FDS already knows it (and
can be prone to errors).

Invoke the methods to compute the odata and UI strings from the FDS,
instead of relying on the filter, which can be more error prone.

Note that this commits breaks CX filters because we cannot implement the
odata and UI string methods yet (we need the following commits for
that).