LPS-202104 Simplify how CET objects are created
Since we have added the notion of raw CET objects in the previous commit
and since we are putting more and more logic into each CET type
implementation I'm moving as much responsibility as possible to the
CETImplFactory classes and removing it from the CET implementation.

This:

1. Simplifies a lot the implementation of new types.
2. Centralizes operations in expected places.
3. And makes it easier to manage CET types without having to create CET
   instances just to ask them.

Note that we can dig deeper into 3 because there is still some code that
needs CET instantions but could be invoked without creating that.

I'll refactor that in further commits.