LPS-202104 Remove the raw CET concept
After the previous refactors, I've realized we don't really need a new
not very well defined concept like raw CET. We can simply move it to a
boolean parameter in the factory and it's the caller's responsibility to
make sure it doesn't use the incorrect CET version (raw vs those with
variables replaced).

The fact is that we don't mix raw and replaced CETs in any part of DXP, we
either use all raw or all replaced CETs, so we can do this new refactor.