Show how different invocation of fixtures return different objects
So, in this commit I request a 'dog' pet two times and I show in the
console a message to see what happens.

As expected, the fixture creates two different dogs (see the i):

----

Running 9 tests using 4 workers
[pets] › pets/dog.spec.js:14:1 › get first dog pet
======> got first dog pet named: { type: 'dog', name: 'Tony', i: 0 }
[pets] › pets/cat.spec.js:14:1 › get a cat pet
======> got cat pet named: { type: 'cat', name: 'Garfield', i: 0 }
[pets] › pets/dog.spec.js:20:1 › get second dog pet
======> got secong dog pet named: { type: 'dog', name: 'Tony', i: 1 }

----

If we were creating random data in the DB this would create two rows which,
from the performance POV, may be quite heavy if tests are too granular.

However, there are situations were we may want that several tests share
the same dependency because we know they won't collide.

Think of tests for Echo team where they test features for "Sites". Chances
are they will have multiple tests that can share one site that is created
at the start of the test run and teared down at the end of the test run (as
opposed to create/destroy in each test).

For that we can use the pooled fixtures technique that I'll demonstrate in
the next commit.