Switch from per-test create/destroy of deps to pooling
I'm pooling one single dog and cat among all tests. In this case I assume
the pets are read-only so there's no problem in sharing them in several
tests that are run in parallel.

However, it seems very easy how we could create a mechanism to make tests
relying on the same fixture dep be serialized automagically by the fixture
or make them fail, or whatever we decide to do in that situation for any
specific fixture dep type and name.