LPD-18820 Clean webpack output
As per
https://github.com/liferay-frontend/liferay-portal/pull/4019#discussion_r1507155428

This is to clean leftovers of previous builds so that they don't end up in
the ZIP file.

Note however that this won't really fix the error because the Workspace is
caching the output files of previous builds even if we clean them in the
npm side so, until that is fixed, the users will need to do a gradle clean
every time they build the project.

See https://liferay.slack.com/archives/CL08WH5V0/p1709281374316849 for more
info.