LPD-19057 Prevent flakyness when a backendPage is used
Apparently if the first page to hit a URL is not the default test page,
sometimes the tests fail to pass and the traces just show the backend page.

I haven't been able to reproduce the bug in a clean Playwright
installation, but after many tests I've done locally it seems we need to
make the first navigation in the test page.