LPD-25772 Fix custom element test
I'm removing custom element 2 since what it really does is demonstrating
how a backend and frontend CX may interact and what we want to see is if
the CX framework works.

Testing that on such a complex example is error prone and doesn't bring any
benefit and, in any case, we are already testing a lot of custom elements.