LPD-40070 Substitute <liferay-util:on> by <liferay-ui:csp>
Instead of creating a JSP tag that attaches an event handler to the previous DOM node, which is
something that can fail due to SPA moving the nodes, refactoring the code, and so on..., we are
creating a tag that rewrites all its body extracting inline handlers to their own <script> node
using `ContentSecurityPolicyHTMLRewriterUtil`.

To enahnce performance, this transformation only takes place when a CSP nonce is active for the
current request.