LPD-47925 Rewrite imports after moving code
Installing a dynamic include to bootstrap Liferay global object moved
funcctions doesn't work due to this:

https://github.com/liferay-frontend/liferay-portal/pull/4720#issuecomment-2647499809

So I'm changing the strategy and moving the initialization of
Liferay methods to top_js.jspf and the typescript declarations to
global.d.ts.

The rationale is that I'll be treating all Liferay.xxx stuff as legacy and
the things in ESM modules as clean code (free from all legacy).

Thus it will be the problem of top_js.jspf and the legacy layer to know how
to load the clean stuff to make things work as before.

My hope is that by moving things to that legacy global layer will make it
easier in the future to delete all legacy stuff without touching anything
at the clean side.