THE ONE-AND-ONLY™️ COMMIT IN THIS BRANCH
This commit contains the GitHub actions workflow definition for the
performance report generation.

The workflow is launched every time changes are pushed to this branch
and assumes that whatever is before this commit is the subject under
analysis, ie: the commit SHA that is shown in the reports will be the
one just before this commit.

For that reason, there should NOT be more commits on top of this one in
this branch (unless we update the workflow definition) because doing so
will generate invalid hashes for the reports and traceability will be
lost.

If anything needs to be changed in the workflow we can simply ammend
this commit and force push this branch.

Note that this branch will be rebased with `master` every day and pushed
so that we can generate daily performance reports and publish them to
https://izaera.github.io/liferay-portal/.